+++ packages/core/src/index.js
@@ -1,6 +1,6 @@
 module.exports = {
   Secrez: require('./Secrez'),
-  Crypto: require('./Crypto'),
+  Crypto: require('@secrez/crypto'),
   Entry: require('./Entry'),
   config: require('./config'),
   ConfigUtils: require('./config/ConfigUtils'),
