+++ packages/crypto/.jshintrc
@@ -1,0 +1,3 @@
+{
+  "esversion": 8
+}
