+++ packages/crypto/.npmignore
@@ -1,0 +1,6 @@
+test
+.nyc_output
+coverage
+tmp
+.idea
+.scratch
