+++ packages/crypto/README.md
@@ -1,0 +1,30 @@
+# @secrez/crypto
+
+Secrez is the secrets manager for the cryptocurrencies era.
+
+@secrez/crypto is the basic crypto library.
+
+## TODO
+
+API documentation
+
+## History
+
+__0.1.0__
+* moved the package from @secrez/core to a separate libr 
+
+
+## Test coverage
+
+```
+ 
+```
+
+
+## Copyright
+
+(c) 2020-present [Francesco Sullo](https://francesco.sullo.co) (<francesco@sullo.co>)
+
+## Licence
+
+MIT
