+++ packages/crypto/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "@secrez/core",
-  "version": "0.8.8",
+  "name": "@secrez/crypto",
+  "version": "0.1.0",
   "license": "MIT",
   "scripts": {
     "lint": "eslint -c .eslintrc 'src/**/*.js' 'test/**/*.js'",
@@ -10,13 +10,8 @@
     "posttest": "nyc check-coverage --statements 99 --branches 95 --functions 99 --lines 99"
   },
   "dependencies": {
-    "@secrez/crypto": "workspace:~0.1.0",
-    "@secrez/utils": "workspace:~0.1.5",
     "base-x": "^3.0.8",
     "bip39": "^3.0.2",
-    "fs-extra": "^8.1.0",
-    "homedir": "^0.6.0",
-    "lodash": "^4.17.15",
     "microtime": "^3.0.0",
     "sha3": "^2.1.2",
     "shamir": "^0.7.1",
