+++ packages/crypto/test/fixtures/index.js
@@ -1,0 +1,40 @@
+module.exports = {
+  password: 'unaSTRANA342',
+  newPassword: 'sa7sau3h33y2hhes',
+  b58Hash: 'J5JDehfYiYFyRcySdHQNiCJdVrsrUyVtpCEXU7fBQ6q3',
+  b32Hash: 'b9poxy4pq63ow7ex8ckksbioyb3o7az1zpjkoozhxdm15ngeryhq',
+  passwordSHA3Hex: 'fdb07834d77b30a750f3b14ab06b000730ee2f2bb52a842fc78d72d88c82038e',
+  passwordB64: 'dW5hU1RSQU5BMzQy',
+  salt: 'someSalt',
+  iterations: 23456,
+  iterationsB58: '7Yq',
+  hash23456iterations: 'GCF7ytpi9DyMPbuDhLj6vW1oSe99nBTLzABcb1qvTeLY',
+  hash23456iterationsNoSalt: 'ErQE64oXPPs7QqbirznqzCap3KNyMKDZZZBexRfhS7Pb',
+  hash23456iterationsNoSaltVersionTwo: '621vXeh5P96cRCjRvEtiquBjQ9nDv5vWqpzENyMoWia5',
+  secondFactor: {
+    authenticator: 'yubo',
+    id: '7sheywt3rfse',
+    salt: 'jskeywhetduw63t4f45stwgeyrhtteur',
+    secret: 'sadsasewiuwnefdnsd09f38092erdu9feaofjfwy2feqoasz',
+    credential: 'sadadada',
+    recoveryCode: 'salad spring peace silk snake real they thunder please final clinic close',
+    wrongMnemonic: 'they thunder please final clinic close salad spring peace silk snake real'
+  },
+  passphrase: 'when John breaks the cronicle',
+  signaturePair: {
+    publicKey: [
+      144, 197, 186, 79, 209, 26, 15, 35,
+      184, 222, 122, 255, 252, 165, 120, 192,
+      81, 98, 40, 50, 66, 191, 67, 46,
+      255, 166, 77, 178, 72, 107, 170, 124
+    ],
+    secretKey: [
+      74, 141, 176, 24, 196, 155, 253, 186, 220, 199, 47,
+      100, 204, 117, 251, 55, 1, 251, 42, 154, 81, 82,
+      166, 60, 222, 71, 146, 10, 240, 198, 57, 226, 144,
+      197, 186, 79, 209, 26, 15, 35, 184, 222, 122, 255,
+      252, 165, 120, 192, 81, 98, 40, 50, 66, 191, 67,
+      46, 255, 166, 77, 178, 72, 107, 170, 124
+    ]
+  }
+}
